% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_difference_by_region.R
\name{plot_difference_by_region}
\alias{plot_difference_by_region}
\title{Difference Plots for 9 ROI}
\usage{
plot_difference_by_region(
  data,
  plotname,
  conditionToPlot = MM_RAW,
  levelA = semMM_RAW,
  levelB = consistent,
  color_palette,
  ant_levels = anteriority_3l,
  med_levels = mediality_a,
  vary = Voltage,
  group_var,
  show_group_obs,
  y_annot = 6,
  delta = 1,
  baseline = c(-500, -200),
  time_windows = list(c(-250, -150), c(-150, 50), c(50, 200), c(200, 300), c(300, 500),
    c(500, 700), c(700, 900)),
  plot_title = " "
)
}
\arguments{
\item{conditionToPlot}{column of the dataframe with the condition to plot}

\item{file}{dataframe containing}
}
\value{
A PDF file containing the Difference plots by region
}
\description{
This function creates a plot file with 9 electrodes of interests displaying ERPs
for different conditions. It will need a loaded dataframe with your EEG data
and a column indicating the condition to display.
It assumes that there is a column named Voltage with your voltage values.
Default values are provided for electrodes but it can be customized.
}
