% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_difference_maps.R
\name{plot_difference_maps}
\alias{plot_difference_maps}
\title{Difference Plots for 9 ROI}
\usage{
plot_difference_maps(
  data,
  plotname = "auto",
  conditionToPlot = MM_RAW,
  levelA = semMM_RAW,
  levelB = consistent,
  output_type = "pdf",
  vary = Voltage,
  group_var = Subject,
  topoplots_data = "voltage_difference",
  topoplots_time_windows = list(c(-250, -150), c(-150, 50), c(50, 200), c(200, 300),
    c(300, 500), c(500, 700), c(700, 900)),
  topoplots_scale = c(-2, 2),
  t_test_threshold = 0.05
)
}
\arguments{
\item{conditionToPlot}{column of the dataframe with the condition to plot}

\item{file}{dataframe containing}
}
\value{
A PDF file containing the Difference plots by region
}
\description{
This function creates a plot file with 9 electrodes of interests displaying ERPs
for different conditions. It will need a loaded dataframe with your EEG data
and a column indicating the condition to display.
It assumes that there is a column named Voltage with your voltage values.
Default values are provided for electrodes but it can be customized.
}
