% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_erp_by_electrode.R
\name{plot_erp_by_electrode}
\alias{plot_erp_by_electrode}
\title{Plot ERP for 9 electrodes of interest}
\usage{
plot_erp_by_electrode(
  data,
  conditionToPlot,
  electrodes_list = c("F3", "Fz", "F4", "C3", "Cz", "C4", "P3", "Pz", "P4"),
  output_type = "pdf",
  color_palette = c("#4DAF4A", "#377EB8", "#FF7F00", "#984EA3", "#000000"),
  baseline = c(-2450, -2250),
  tick_distance = 200,
  vary = "Voltage",
  plotname = "auto",
  check_message = FALSE,
  conf_interval = FALSE,
  y_annot = 6,
  delta = 1,
  rectangles = list(list(-2250, -1600, "N1"), list(-1500, -850, "BA/BEI"), list(-750,
    -100, "N2"), list(0, 650, "Verb"))
)
}
\arguments{
\item{conditionToPlot}{column of the dataframe witht the condition to plot}

\item{file}{dataframe containing}
}
\value{
A PDF file containing the ERP plots
}
\description{
This function creates a plot file with 9 electrodes of interests displaying ERPs
for different conditions. It will need a loaded dataframe with your EEG data
and a column indicating the condition to display.
It assumes that there is a column named Voltage with your voltage values.
Default values are provided for electrodes but it can be customized.
}
