% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_erp_by_region.R
\name{plot_erp_by_region}
\alias{plot_erp_by_region}
\title{Plot ERP by region of interest}
\usage{
plot_erp_by_region(
  data,
  conditionToPlot,
  plotname = "auto",
  output_type = "pdf",
  color_palette = c("#4DAF4A", "#377EB8", "#FF7F00", "#984EA3", "#000000"),
  ant_levels = "anteriority_3l",
  med_levels = "mediality_k",
  vary = "Voltage",
  y_annot = 6,
  delta = 1,
  baseline = c(-500, -200)
)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function creates a plot file with 9 regions of interests displaying ERPs
for different conditions indicated in the input.
}
