% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-priming-data.R
\docType{data}
\name{relpriming}
\alias{relpriming}
\title{Herbay, Royle and Steinhauer data on relational priming}
\format{
An object of class \code{data.frame} with 799986 rows and 7 columns.
}
\usage{
data(relpriming)
}
\description{
Data from a nERP experiment on relational priming, with data on 11 participants and 3 conditions
The outcome is the EEG between -650ms and 900ms
}
\examples{
data(relpriming)

\donttest{
iplotCurves(phe, times, phe[,c(61,121)], phe[,c(121,181)],
            chartOpts=list(curves_xlab="Time (hours)", curves_ylab="Root tip angle (degrees)",
                           scat1_xlab="Angle at 2 hrs", scat1_ylab="Angle at 4 hrs",
                           scat2_xlab="Angle at 4 hrs", scat2_ylab="Angle at 6 hrs"))}
}
\references{
Herbay AH, Royle P, Steinhauer K,
(in preparation) Relational Priming POP2 title
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
