% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ERP_stats_table.R
\name{generate_ERP_stats_table}
\alias{generate_ERP_stats_table}
\title{Generate color coded stats table for ERPs}
\usage{
generate_ERP_stats_table(
  data,
  model_structure,
  output_name,
  timeWindowMode,
  model_type = "lm",
  custom_TW = list(c(-300, -150), c(-150, 50), c(50, 200), c(200, 300), c(300, 500),
    c(500, 700), c(700, 900)),
  time_step,
  min_time,
  max_time
)
}
\arguments{
\item{data}{dataframe containing eeg data}

\item{model_structure}{string with the lmer model}

\item{output_name}{string for the output name}
}
\value{
An HTML file containing thes stats tables
}
\description{
Description to come
}
