% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_difference.R
\name{plot_difference}
\alias{plot_difference}
\title{Difference Plots for 9 ROI}
\usage{
plot_difference(
  data,
  plotname = "auto",
  conditionToPlot = MM_RAW,
  levelA = semMM_RAW,
  levelB = consistent,
  color_palette,
  output_type = "pdf",
  ant_levels = anteriority_3l,
  med_levels = mediality_a,
  vary = Voltage,
  group_var = Subject,
  show_group_obs = FALSE,
  labels_vertical_position = "auto",
  labels_height = "auto",
  baseline = c(-500, -200),
  topoplots_time_windows = list(c(-250, -150), c(-150, 50), c(50, 200), c(200, 300),
    c(300, 500), c(500, 700), c(700, 900)),
  topoplots_scale = c(-2, 2),
  time_labels_interval = 200,
  custom_labels = list(),
  electrodes_to_display = c()
)
}
\arguments{
\item{conditionToPlot}{column of the dataframe with the condition to plot}

\item{file}{dataframe containing}
}
\value{
A PDF file containing the Difference plots by region
}
\description{
This function creates a plot file with 9 electrodes of interests displaying ERPs
for different conditions. It will need a loaded dataframe with your EEG data
and a column indicating the condition to display.
It assumes that there is a column named Voltage with your voltage values.
Default values are provided for electrodes but it can be customized.
}
