% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_erp.R
\name{plot_erp}
\alias{plot_erp}
\title{Plot ERP for 9 electrodes of interest}
\usage{
plot_erp(
  data,
  conditionToPlot,
  electrodes_list = c("F3", "Fz", "F4", "C3", "Cz", "C4", "P3", "Pz", "P4"),
  baseline = c(-2450, -2250),
  color_palette = c("#4DAF4A", "#EA2721", "#377EB8", "#FF7F00", "#984EA3", "#000000",
    "#5c5c5c", "#945D25", "#FF748C", "#2E692C"),
  output_type = "pdf",
  adjusted_baseline = FALSE,
  time_labels_interval = 200,
  plotname = "auto",
  show_check_message = FALSE,
  show_conf_interval = FALSE,
  custom_labels = list(),
  labels_vertical_position = "auto",
  labels_height = "auto",
  vary = "Voltage"
)
}
\arguments{
\item{conditionToPlot}{column of the dataframe witht the condition and levels to plot}

\item{electrodes_list}{vector of Electrodes names to plot (between quotes)}

\item{file}{dataframe containing eeg data}
}
\value{
A PDF file containing the ERP plots
}
\description{
This function creates a plot file with 9 electrodes of interests displaying ERPs
for different conditions. It will need a loaded dataframe with your EEG data
and a column indicating the condition to display.
It assumes that there is a column named Voltage with your voltage values.
Default values are provided for electrodes but it can be customized.
}
