% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_erp_by_region.R
\name{plot_erp_by_region}
\alias{plot_erp_by_region}
\title{Plot ERP by region of interest}
\usage{
plot_erp_by_region(
  data,
  plotname,
  conditionToPlot,
  color_palette,
  ant_levels = "anteriority_3l",
  med_levels = "mediality_k",
  vary = "Voltage",
  y_annot = 6,
  delta = 1,
  baseline = c(-500, -200)
)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function creates a plot file with 9 regions of interests displaying ERPs
for different conditions indicated in the input.
}
\details{
It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
